+++ src/core/bot.ts
@@ -11,7 +11,7 @@ import { CustomClient } from "../types/discord.js.js";
 import { CommandHandler } from "./commands/commandHandler";
 import { MongoConnector } from "./database";
 import { EventHandler } from "./events/eventHandler";
-import { InteractionHandler } from "./interactions/interactionHandler.js";
+import { InteractionHandler } from "./interactions/InteractionHandler.js";
 import { JobHandler } from "./jobs/jobHandler";
 import { LogHandler } from "./log";
 config();
